pyproject.toml
[project]
name = "mcp-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.0.2",
    "mcp-use>=1.4.0",
    "mcp[cli]>=1.20.0",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]
